requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio
mlflow
jsonargparse[signatures]
pytorch-lightning[extra]
ipykernel
onnxruntime-gpu
inference-schema
ipykernel
fibberio
mltable
sklearn
python-dotenv
opencensus-ext-azure